Bing API map image downloads based off of player location and live player location on map
By far the biggest commit into this project. This marks the date where all components of the object are now actually working in the game. From here on out it'll be just be plug and play of the different components that are already functioning. Hell Yeah!!!

Note: I commit this at like 5 am so I couldnt go out and test out if the location functionality works but if the minor error(s) does arise, I will be fixed on next commit.